requirements.txt
Django==4.2.3
django-cors-headers==4.2.0
django-debug-toolbar==4.2.0
django-enumfields==2.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
mysqlclient==2.2.0
Pillow==10.0.0
drf-yasg==1.21.7